+++ bugsnag-android-core/src/test/java/com/bugsnag/android/SessionTrackerPauseResumeTest.kt
@@ -48,6 +48,7 @@ internal class SessionTrackerPauseResumeTest {
 
     @Before
     fun setUp() {
+        `when`(client.getNotifier()).thenReturn(Notifier())
         `when`(client.getAppContext()).thenReturn(context)
         `when`(client.getAppDataCollector()).thenReturn(appDataCollector)
         `when`(appDataCollector.generateApp()).thenReturn(app)